Seção 17 - Palestrantes e Redes Sociais - Frontend

+++ Back/src/ProEventos.API/Helpers/Util.cs
@@ -38,10 +38,12 @@ namespace ProEventos.API.Helpers
 
         public void DeleteImage(string imageName, string destino)
         {
-            var imagePath = Path.Combine(_hostEnvironment.ContentRootPath,  @$"Resources/{destino}", imageName);
-            if (System.IO.File.Exists(imagePath))
-            {
-                System.IO.File.Delete(imagePath);
+            if(!string.IsNullOrEmpty(imageName)){
+                var imagePath = Path.Combine(_hostEnvironment.ContentRootPath,  @$"Resources/{destino}", imageName);
+                if (System.IO.File.Exists(imagePath))
+                {
+                    System.IO.File.Delete(imagePath);
+                }
             }
         }
     }
